Move markdown to ui module
