+++ ui/src/main/java/com/neaniesoft/vermilion/ui/markdown/CommonMarkModule.kt
@@ -1,4 +1,4 @@
-package com.neaniesoft.vermilion.postdetails.data
+package com.neaniesoft.vermilion.ui.markdown
 
 import dagger.Module
 import dagger.Provides
